Move all results fetching to facets and data components.
